Refactor ImageViewerActivity and ImagePagerAdapter
