+++ app/src/main/java/io/github/tjg1/nori/ImageViewerActivity.java
@@ -81,6 +81,10 @@ public class ImageViewerActivity extends AppCompatActivity implements ViewPager.
   /** True if the {@link AppBarLayout} is currently collapsed. */
   private boolean appBarCollapsed = false;
 
+  public SearchResult getSearchResult() {
+    return searchResult;
+  }
+
   @Override
   protected void onCreate(Bundle savedInstanceState) {
     // Restore state from savedInstanceState.
@@ -135,7 +139,7 @@ public class ImageViewerActivity extends AppCompatActivity implements ViewPager.
     }
 
     // Create and set the image viewer Fragment pager adapter.
-    imagePagerAdapter = new ImagePagerAdapter(getSupportFragmentManager());
+    imagePagerAdapter = new ImagePagerAdapter(getSupportFragmentManager(),this);
     viewPager = (ImageViewerPager) findViewById(R.id.image_pager);
     viewPager.setAdapter(imagePagerAdapter);
     viewPager.addOnPageChangeListener(this);
@@ -327,55 +331,6 @@ public class ImageViewerActivity extends AppCompatActivity implements ViewPager.
     return request;
   }
 
-  /** Adapter used to populate {@link android.support.v4.view.ViewPager} with {@link io.github.tjg1.nori.fragment.ImageFragment}s. */
-  private class ImagePagerAdapter extends FragmentStatePagerAdapter {
-    private ImageFragment activeFragment;
-
-    public ImagePagerAdapter(FragmentManager fm) {
-      super(fm);
-    }
-
-    @Override
-    public Fragment getItem(int position) {
-      // Create a new instance of ImageFragment for the given image.
-      Image image = searchResult.getImages()[position];
-
-      if (shouldUseVideoPlayerFragment(image)) {
-        return VideoPlayerFragment.newInstance(image);
-      } else {
-        return RemoteImageFragment.newInstance(image);
-      }
-    }
-
-    @Override
-    public void setPrimaryItem(ViewGroup container, int position, Object object) {
-      super.setPrimaryItem(container, position, object);
-      if (activeFragment != object) {
-        if (activeFragment != null) {
-          activeFragment.onHidden();
-        }
-        activeFragment = (ImageFragment) object;
-        activeFragment.onShown();
-      }
-    }
-
-    @Override
-    public int getCount() {
-      // Return the search result count.
-      if (searchResult == null) {
-        return 0;
-      }
-      return searchResult.getImages().length;
-    }
-
-    /** Returns true if the {@link Image} object is a WebM/MP4 animation. */
-    private boolean shouldUseVideoPlayerFragment(Image image) {
-      String path = Uri.parse(image.fileUrl).getPath();
-      String fileExt = path.contains(".") ? path.toLowerCase(Locale.US)
-          .substring(path.lastIndexOf(".") + 1) : null;
-      return "mp4".equals(fileExt) || "webm".equals(fileExt);
-    }
-  }
 
   /** Callback waiting to receive more images for infinite scrolling. */
   private class InfiniteScrollingSearchCallback implements SearchClient.SearchCallback {
